+++ quickstarts/azure-openai-assistants/ts/index-using-password.ts
