+++ quickstarts/azure-openai-assistants/ts/index.ts
