+++ quickstarts/azure-openai-assistants/ts/package.json
@@ -7,19 +7,17 @@
   "scripts": {
     "create-resource": "bash scripts/create-azure-resource.sh",
     "build": "tsc",
-    "start": "node dist/index.js",
-    "test": "ts-node"
+    "start": "node index.js",
+    "dev": "node -r dotenv/config index.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "@azure/identity": "^4.0.1",
-    "openai": "^4.42.0",
-    "dotenv": "^16.4.5"
+    "openai": "^4.42.0"
   },
   "devDependencies": {
     "@types/node": "^20.12.2",
-    "ts-node": "^10.9.2",
     "typescript": "^5.4.3"
   }
 }
